Correct serious flaw in plot.py
There was a serious flaw in plot.py as originally written (January
22, 202O). The mean used to compute standard deviation (z-score)
coordinates averaged the frequencies of "in" and "non" in the R1
and R2 dicta as if the samples were the same size and had equal
weight. This version has been corrected to use the mean frequencies
of "in" and "non" over the total population (R1 + R2) to compute
z-scores.