Change calculation of mean
Changing the calculation of mean again. The mean has to be calculated
with respect to all the samples (Gratian[012].txt and dePen[12].txt)
Calculating the standard deviation distance of the R1 and R2 dicta
away from a mean calculated on the basis of the R1 + R2 dicta only
measures the size difference between the samples.